prune: Remove deployer pods when pruning failed deployments
