+++ pkg/deploy/prune/resolvers_test.go
@@ -11,6 +11,7 @@ import (
 	"k8s.io/kubernetes/pkg/util/sets"
 
 	deployapi "github.com/openshift/origin/pkg/deploy/api"
+	deployutil "github.com/openshift/origin/pkg/deploy/util"
 )
 
 type mockResolver struct {
@@ -150,8 +151,8 @@ func TestPerDeploymentConfigResolver(t *testing.T) {
 					failedDeployments = append(failedDeployments, deployment)
 				}
 			}
-			sort.Sort(sortableReplicationControllers(completedDeployments))
-			sort.Sort(sortableReplicationControllers(failedDeployments))
+			sort.Sort(deployutil.ByMostRecent(completedDeployments))
+			sort.Sort(deployutil.ByMostRecent(failedDeployments))
 			purgeCompleted := []*kapi.ReplicationController{}
 			purgeFailed := []*kapi.ReplicationController{}
 			if keep >= 0 && keep < len(completedDeployments) {
